requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-microservice-request==0.5.3
django-ninja==0.22.2
django-polymorphic==3.1.0
djangorestframework==3.14.0
idna==3.4
pydantic==1.10.9
python-dotenv==1.0.0
django-microservice-request==0.5.3
django-ninja==0.22.2
django-polymorphic==3.1.0
djangorestframework==3.14.0
idna==3.4
pydantic==1.10.9
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
gevent==23.7.0
greenlet==2.0.2
gunicorn==21.2.0
psycopg2-binary==2.9.6